factor out filter settings
